separate animation & algo logic
